requirements.txt
django>=4.0.0,< 4.1.0
djangorestframework>=3.13.1,<3.14
psycopg2>=2.6.8,<2.9
openpyxl
coverage
#kafka-python
kazoo
redis
django-redis
# celery[kafka]
celery[redis]>=4.7.7
kombu>=5.0.2
django-debug-toolbar
flower